lib: utils/fdt: add fdt_device_add_memrange_to_root_domain
When many devices need pmp protection, and their memory is not contiguous,
if each device occupies one pmp, there will not be enough pmps available.
We can create a large block of memregion, and create a "regions" on their
DT node pointing to this large block of memregion. Such multiple devices
share a memregion.

The fdt_device_add_memrange_to_root_domain function analyzes the DT node
of the device and adds the memregion to root domain.

Signed-off-by: Xiang W <wxjstz@126.com>